pyproject.toml
[tool.poetry]
name = "video-lessons-django"
version = "0.1.0"
description = ""
authors = ["khage"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
djangorestframework = "^3.14.0"
envparse = "^0.2.0"
opencv-python = "^4.8.0.76"
drf-spectacular = "^0.26.5"
django-filter = "^23.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
